+++ NEWS.md
@@ -5,6 +5,8 @@
   colums (could happen when CID/HCD and UVPD scans are taken independently with
   different software versions) [2017-12-22].
 - Don't replace NA values with zeros in the `colData` [2017-12-22].
+- Convert On/Off `character` columns in scan and method information to
+  `logical` [2017-12-22].
 
 ## Changes in version 1.1.1
 - Respect assigned intensity in conditions for `bestConditions,NCBSet-method`